More file position data type changes. Also reflecting this in user code.
git-svn-id: https://irrlicht.svn.sourceforge.net/svnroot/irrlicht/trunk@954 dfc29bdd-3216-0410-991c-e03cc46cb475

+++ include/irrXML.h
@@ -224,7 +224,7 @@ namespace io
 		virtual int read(void* buffer, int sizeToRead) = 0;
 
 		//! Returns size of file in bytes
-		virtual int getSize() = 0;
+		virtual long getSize() = 0;
 	};
 
 	//! Empty class to be used as parent class for IrrXMLReader.
